Refactor project generator to use template files
