+++ generator_templates/README.md
@@ -1,0 +1,6 @@
+# {nom_site}
+
+Ce site est généré automatiquement.
+
+- Couleur principale : {couleur}
+- Dossier : {slug}
