+++ generator_templates/cgu.html
@@ -1,0 +1,2 @@
+<h1>Conditions Générales d'Utilisation</h1>
+<p>Bienvenue sur le site {nom_site}. Les présentes conditions régissent l'utilisation du service proposé.</p>
