+++ generator_templates/footer.html
@@ -1,0 +1,9 @@
+<footer>
+    <p>&copy; {{ nom_site }} - Tous droits réservés</p>
+    <nav>
+        <a href="{{ url_for('cgu') }}">CGU</a> |
+        <a href="{{ url_for('mentions_legales') }}">Mentions légales</a> |
+        <a href="{{ url_for('politique_confidentialite') }}">Confidentialité</a> |
+        <a href="{{ url_for('parametres_cookies') }}">Cookies</a>
+    </nav>
+</footer>
