+++ generator_templates/mentions-legales.html
@@ -1,0 +1,3 @@
+<h1>Mentions légales</h1>
+<p><strong>Nom du site :</strong> {nom_site}</p>
+<p><strong>Email :</strong> contact@exemple.com</p>
