+++ generator_templates/parametres-cookies.html
@@ -1,0 +1,2 @@
+<h1>Paramètres des cookies</h1>
+<p>Ce site utilise des cookies pour améliorer votre expérience.</p>
