+++ generator_templates/politique-confidentialite.html
@@ -1,0 +1,2 @@
+<h1>Politique de confidentialité</h1>
+<p>Cette politique explique comment {nom_site} collecte et utilise vos données personnelles.</p>
