+++ generator_templates/requirements.txt
@@ -1,0 +1,1 @@
+Flask
