+++ generator_templates/style.css
@@ -1,0 +1,2 @@
+body { font-family: Arial; margin: 0; padding: 0; }
+header { background-color: {couleur}; color: white; padding: 10px 20px; display: flex; align-items: center; gap: 20px; }
